requirements.txt
websockets>=12.0,<13.0
httpx>=0.27.0,<1.0
pydantic>=2.5,<3.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=10.0
playwright>=1.40,<2.0
customtkinter>=5.2,<6.0
pyinstaller>=6.0,<7.0
